requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.11
chromadb==0.4.18
tiktoken==0.5.2
pydantic==2.5.2
python-dotenv==1.0.0
pypdf==3.17.1
pandas==2.1.3
numpy==1.26.2
pytest==7.4.3

